Create Portfolio Website
Project 2)

Basic Portfolio Website : 
My Portfolio Website , The portfolio website is a single-page application that displays my projects, social media handles, and experience. The website is designed to be responsive, meaning it adapts to different screen sizes and devices. The website has a clean and minimalistic design, with a focus on showcasing my projects and skills.